refactoring of graphql

+++ graphql/src/data/reports.ts
@@ -1,6 +1,9 @@
 import { collection } from "../mongo";
 import { Resolver, ID } from "../Types";
 import uuid from "uuid/v4";
+import Dataloader from "dataloader";
+import { groupBy, keyBy } from "lodash";
+
 const run = collection("reports");
 
 export type Report = {
@@ -12,29 +15,25 @@ export type Report = {
   owner: ID;
 };
 
-type Slide = {
+export type Slide = {
   id: ID;
-  createdAt: Date;
-  updatedAt: Date;
   elements: SlideElement[];
 };
 
 type ReportUpdate = {
   id: ID;
-  title: string;
-  slides: Slide[];
+  title?: string;
+  slides?: Slide[];
+  owner?: ID;
 };
 
 type ReportCreate = {
   title: string;
-  organisation: ID;
-  slides: Slide[];
+  owner: ID;
 };
 
 export type SlideElement = {
   id: ID;
-  createdAt: Date;
-  updatedAt: Date;
   type: "TEXT" | "CHART";
   x: number;
   y: number;
@@ -51,36 +50,38 @@ type SlideText = {
 };
 
 export type SlideChart = {
-  dataStore: ID;
+  datastore: ID;
   query: string;
   xAxis: string;
   yAxis: string;
 };
 
-export const fetchAll: Resolver<Report[]> = (parent, args, { user }) => {
-  console.log("reports");
-  return run((reports) => reports.find({ owner: user!.id }).toArray());
-};
+const ownerLoader = new Dataloader<string, Report[]>(async (owners) => {
+  const reports = await run<Report[]>((reports) =>
+    reports.find({ owner: { $in: owners } }).toArray()
+  );
+  const indexed = groupBy(reports, (report) => report.owner);
+  return owners.map((owner) => indexed[owner] || []);
+});
 
-export const fetch: Resolver<Report | null, { id: ID }> = (
-  parent,
-  { id },
-  { user }
-) => {
-  return run((reports) => reports.findOne({ _id: id, owner: user!.id }));
-};
+export const listByOwner = ownerLoader.load;
 
-export const create: Resolver<Report | null, { title: string }> = async (
-  parent,
-  { title },
-  { user }
-) => {
+const idLoader = new Dataloader<string, Report | undefined>(async (ids) => {
+  const reports = await run<Report[]>((reports) =>
+    reports.find({ _id: { $in: ids } }).toArray()
+  );
+  const indexed = keyBy(reports, (report) => report._id);
+  return ids.map((id) => indexed[id]);
+});
+
+export const get = idLoader.load;
+
+export const create = async (report: ReportCreate) => {
   const createdAt = new Date();
   const updatedAt = new Date();
   const newReport: Report = {
+    ...report,
     _id: uuid(),
-    title,
-    owner: user!.id,
     createdAt,
     updatedAt,
     slides: []
@@ -90,25 +91,17 @@ export const create: Resolver<Report | null, { title: string }> = async (
   return newReport;
 };
 
-export const remove: Resolver<ID, { id: ID }> = async (
-  parent,
-  { id },
-  { user }
-) => {
+export const remove = async (id: string) => {
   await run((reports) => reports.deleteOne({ _id: id }));
   return id;
 };
 
-export const removeAll: Resolver<number | null> = async () => {
-  let res = await run((reports) => reports.deleteMany({}));
+export const removeAll = async (owner: ID) => {
+  let res = await run((reports) => reports.deleteMany({ owner }));
   return res.result.ok ? (res.result.n as number) : null;
 };
 
-export const update: Resolver<Report | null, { report: ReportUpdate }> = async (
-  parent,
-  { report },
-  { user }
-) => {
+export const update = async (report: ReportUpdate) => {
   await run((reports) =>
     reports.updateOne({ _id: report.id }, { $set: report })
   );
